fix small edge case, add more tests, refactor
