+++ src/function/matrix/eigs/realSymetric.js
@@ -8,10 +8,6 @@ export function createRealSymmetric ({ config, addScalar, subtract, abs, atan, c
    * @param {'number' | 'BigNumber'} type
    */
   function main (arr, N, prec = config.epsilon, type) {
-    if (type === 'number' && N === 2) {
-      return trivialCase(arr)
-    }
-
     if (type === 'number') {
       return diag(arr, prec)
     }
@@ -21,32 +17,6 @@ export function createRealSymmetric ({ config, addScalar, subtract, abs, atan, c
     }
 
     throw TypeError('Unsupported data type: ' + type)
-  }
-
-  function trivialCase (arr) {
-    const a = arr[0][0]
-    const b = arr[0][1]
-    const c = arr[1][0]
-    const d = arr[1][1]
-    const values = eigenvalues2x2(a, b, c, d)
-
-    const eigenvector1 = [divideScalar(c, subtract(values[1], a)), 1]
-    const eigenvector2 = [divideScalar(c, subtract(values[0], a)), 1]
-
-    return {
-      values,
-      vectors: [eigenvector1, eigenvector2]
-    }
-  }
-
-  function eigenvalues2x2 (a, b, c, d) {
-    // λ± = ½ trA ± ½ √( tr²A - 4 detA )
-    const trA = addScalar(a, d)
-    const detA = subtract(multiplyScalar(a, d), multiplyScalar(b, c))
-    const x = multiplyScalar(trA, 0.5)
-    const y = multiplyScalar(sqrt(subtract(multiplyScalar(trA, trA), multiplyScalar(4, detA))), 0.5)
-
-    return [subtract(x, y), addScalar(x, y)]
   }
 
   // diagonalization implementation for number (efficient)
